moving opcodes to appropriate sources
